Add toplevel src folder and refactor under it
The Eclipse automation for Gradle or Maven natures was not working so I am trying to create a standard structure with a top level src folder with packages under that.
