+++ Client-Server/bin/.gitignore
@@ -1,0 +1,5 @@
+/severMain/
+/serverMain/
+/ioFormat/
+/dbConnect/
+/exampleClientServer/
