--- Client-Server/business/.gitignore
@@ -1,3 +1,0 @@
-/MainBusiness.class
-/UserFields.class
-/Console.class
