--- Client-Server/dataConnection/.gitignore
@@ -1,1 +1,0 @@
-/DataConnection.class
