--- Client-Server/io/server/.gitignore
@@ -1,1 +1,0 @@
-/UserJsonFile.class
