+++ Client-Server/src/dbConnect/DataConnection.java
@@ -1,4 +1,4 @@
-package dataConnection;
+package dbConnect;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
