+++ Client-Server/src/serverMain/UserFields.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package business;
+package serverMain;
 
 //import java.time.LocalDateTime;
 //import com.google.gson.Gson;
