--- Client-Server/testClientServer/.gitignore
@@ -1,2 +1,0 @@
-/TestClient.class
-/TestServer.class
